+++ src/main/java/com/jsstack/javastack/model/FingerPrint2x2.java
@@ -10,18 +10,18 @@ public class FingerPrint2x2 implements Serializable {
 
 	private int id;
 	private int imageId;
-	private int h1;
-	private int s1;
-	private int v1;
-	private int h2;
-	private int s2;
-	private int v2;
-	private int h3;
-	private int s3;
-	private int v3;
-	private int h4;
-	private int s4;
-	private int v4;
+	private float h1;
+	private float s1;
+	private float v1;
+	private float h2;
+	private float s2;
+	private float v2;
+	private float h3;
+	private float s3;
+	private float v3;
+	private float h4;
+	private float s4;
+	private float v4;
 
 	public int getId() {
 		return id;
@@ -39,99 +39,99 @@ public class FingerPrint2x2 implements Serializable {
 		this.imageId = imageId;
 	}
 
-	public int getH1() {
+	public float getH1() {
 		return h1;
 	}
 
-	public void setH1(int h1) {
+	public void setH1(float h1) {
 		this.h1 = h1;
 	}
 
-	public int getS1() {
+	public float getS1() {
 		return s1;
 	}
 
-	public void setS1(int s1) {
+	public void setS1(float s1) {
 		this.s1 = s1;
 	}
 
-	public int getV1() {
+	public float getV1() {
 		return v1;
 	}
 
-	public void setV1(int v1) {
+	public void setV1(float v1) {
 		this.v1 = v1;
 	}
 
-	public int getH2() {
+	public float getH2() {
 		return h2;
 	}
 
-	public void setH2(int h2) {
+	public void setH2(float h2) {
 		this.h2 = h2;
 	}
 
-	public int getS2() {
+	public float getS2() {
 		return s2;
 	}
 
-	public void setS2(int s2) {
+	public void setS2(float s2) {
 		this.s2 = s2;
 	}
 
-	public int getV2() {
+	public float getV2() {
 		return v2;
 	}
 
-	public void setV2(int v2) {
+	public void setV2(float v2) {
 		this.v2 = v2;
 	}
 
-	public int getH3() {
+	public float getH3() {
 		return h3;
 	}
 
-	public void setH3(int h3) {
+	public void setH3(float h3) {
 		this.h3 = h3;
 	}
 
-	public int getS3() {
+	public float getS3() {
 		return s3;
 	}
 
-	public void setS3(int s3) {
+	public void setS3(float s3) {
 		this.s3 = s3;
 	}
 
-	public int getV3() {
+	public float getV3() {
 		return v3;
 	}
 
-	public void setV3(int v3) {
+	public void setV3(float v3) {
 		this.v3 = v3;
 	}
 
-	public int getH4() {
+	public float getH4() {
 		return h4;
 	}
 
-	public void setH4(int h4) {
+	public void setH4(float h4) {
 		this.h4 = h4;
 	}
 
-	public int getS4() {
+	public float getS4() {
 		return s4;
 	}
 
-	public void setS4(int s4) {
+	public void setS4(float s4) {
 		this.s4 = s4;
 	}
 
-	public int getV4() {
+	public float getV4() {
 		return v4;
 	}
 
-	public void setV4(int v4) {
+	public void setV4(float v4) {
 		this.v4 = v4;
 	}
