+++ src/main/java/com/jsstack/javastack/model/FingerPrint3x3.java
@@ -10,33 +10,33 @@ public class FingerPrint3x3 implements Serializable {
 
 	private int id;
 	private int imageId;
-	private int h1;
-	private int s1;
-	private int v1;
-	private int h2;
-	private int s2;
-	private int v2;
-	private int h3;
-	private int s3;
-	private int v3;
-	private int h4;
-	private int s4;
-	private int v4;
-	private int h5;
-	private int s5;
-	private int v5;
-	private int h6;
-	private int s6;
-	private int v6;
-	private int h7;
-	private int s7;
-	private int v7;
-	private int h8;
-	private int s8;
-	private int v8;
-	private int h9;
-	private int s9;
-	private int v9;
+	private float h1;
+	private float s1;
+	private float v1;
+	private float h2;
+	private float s2;
+	private float v2;
+	private float h3;
+	private float s3;
+	private float v3;
+	private float h4;
+	private float s4;
+	private float v4;
+	private float h5;
+	private float s5;
+	private float v5;
+	private float h6;
+	private float s6;
+	private float v6;
+	private float h7;
+	private float s7;
+	private float v7;
+	private float h8;
+	private float s8;
+	private float v8;
+	private float h9;
+	private float s9;
+	private float v9;
 
 	public int getId() {
 		return id;
@@ -54,219 +54,219 @@ public class FingerPrint3x3 implements Serializable {
 		this.imageId = imageId;
 	}
 
-	public int getH1() {
+	public float getH1() {
 		return h1;
 	}
 
-	public void setH1(int h1) {
+	public void setH1(float h1) {
 		this.h1 = h1;
 	}
 
-	public int getS1() {
+	public float getS1() {
 		return s1;
 	}
 
-	public void setS1(int s1) {
+	public void setS1(float s1) {
 		this.s1 = s1;
 	}
 
-	public int getV1() {
+	public float getV1() {
 		return v1;
 	}
 
-	public void setV1(int v1) {
+	public void setV1(float v1) {
 		this.v1 = v1;
 	}
 
-	public int getH2() {
+	public float getH2() {
 		return h2;
 	}
 
-	public void setH2(int h2) {
+	public void setH2(float h2) {
 		this.h2 = h2;
 	}
 
-	public int getS2() {
+	public float getS2() {
 		return s2;
 	}
 
-	public void setS2(int s2) {
+	public void setS2(float s2) {
 		this.s2 = s2;
 	}
 
-	public int getV2() {
+	public float getV2() {
 		return v2;
 	}
 
-	public void setV2(int v2) {
+	public void setV2(float v2) {
 		this.v2 = v2;
 	}
 
-	public int getH3() {
+	public float getH3() {
 		return h3;
 	}
 
-	public void setH3(int h3) {
+	public void setH3(float h3) {
 		this.h3 = h3;
 	}
 
-	public int getS3() {
+	public float getS3() {
 		return s3;
 	}
 
-	public void setS3(int s3) {
+	public void setS3(float s3) {
 		this.s3 = s3;
 	}
 
-	public int getV3() {
+	public float getV3() {
 		return v3;
 	}
 
-	public void setV3(int v3) {
+	public void setV3(float v3) {
 		this.v3 = v3;
 	}
 
-	public int getH4() {
+	public float getH4() {
 		return h4;
 	}
 
-	public void setH4(int h4) {
+	public void setH4(float h4) {
 		this.h4 = h4;
 	}
 
-	public int getS4() {
+	public float getS4() {
 		return s4;
 	}
 
-	public void setS4(int s4) {
+	public void setS4(float s4) {
 		this.s4 = s4;
 	}
 
-	public int getV4() {
+	public float getV4() {
 		return v4;
 	}
 
-	public void setV4(int v4) {
+	public void setV4(float v4) {
 		this.v4 = v4;
 	}
 
-	public int getH5() {
+	public float getH5() {
 		return h5;
 	}
 
-	public void setH5(int h5) {
+	public void setH5(float h5) {
 		this.h5 = h5;
 	}
 
-	public int getS5() {
+	public float getS5() {
 		return s5;
 	}
 
-	public void setS5(int s5) {
+	public void setS5(float s5) {
 		this.s5 = s5;
 	}
 
-	public int getV5() {
+	public float getV5() {
 		return v5;
 	}
 
-	public void setV5(int v5) {
+	public void setV5(float v5) {
 		this.v5 = v5;
 	}
 
-	public int getH6() {
+	public float getH6() {
 		return h6;
 	}
 
-	public void setH6(int h6) {
+	public void setH6(float h6) {
 		this.h6 = h6;
 	}
 
-	public int getS6() {
+	public float getS6() {
 		return s6;
 	}
 
-	public void setS6(int s6) {
+	public void setS6(float s6) {
 		this.s6 = s6;
 	}
 
-	public int getV6() {
+	public float getV6() {
 		return v6;
 	}
 
-	public void setV6(int v6) {
+	public void setV6(float v6) {
 		this.v6 = v6;
 	}
 
-	public int getH7() {
+	public float getH7() {
 		return h7;
 	}
 
-	public void setH7(int h7) {
+	public void setH7(float h7) {
 		this.h7 = h7;
 	}
 
-	public int getS7() {
+	public float getS7() {
 		return s7;
 	}
 
-	public void setS7(int s7) {
+	public void setS7(float s7) {
 		this.s7 = s7;
 	}
 
-	public int getV7() {
+	public float getV7() {
 		return v7;
 	}
 
-	public void setV7(int v7) {
+	public void setV7(float v7) {
 		this.v7 = v7;
 	}
 
-	public int getH8() {
+	public float getH8() {
 		return h8;
 	}
 
-	public void setH8(int h8) {
+	public void setH8(float h8) {
 		this.h8 = h8;
 	}
 
-	public int getS8() {
+	public float getS8() {
 		return s8;
 	}
 
-	public void setS8(int s8) {
+	public void setS8(float s8) {
 		this.s8 = s8;
 	}
 
-	public int getV8() {
+	public float getV8() {
 		return v8;
 	}
 
-	public void setV8(int v8) {
+	public void setV8(float v8) {
 		this.v8 = v8;
 	}
 
-	public int getH9() {
+	public float getH9() {
 		return h9;
 	}
 
-	public void setH9(int h9) {
+	public void setH9(float h9) {
 		this.h9 = h9;
 	}
 
-	public int getS9() {
+	public float getS9() {
 		return s9;
 	}
 
-	public void setS9(int s9) {
+	public void setS9(float s9) {
 		this.s9 = s9;
 	}
 
-	public int getV9() {
+	public float getV9() {
 		return v9;
 	}
 
-	public void setV9(int v9) {
+	public void setV9(float v9) {
 		this.v9 = v9;
 	}
