+++ src/main/java/com/jsstack/javastack/model/FingerPrint4x4.java
@@ -10,54 +10,54 @@ public class FingerPrint4x4 implements Serializable {
 
 	private int id;
 	private int imageId;
-	private int h1;
-	private int s1;
-	private int v1;
-	private int h2;
-	private int s2;
-	private int v2;
-	private int h3;
-	private int s3;
-	private int v3;
-	private int h4;
-	private int s4;
-	private int v4;
-	private int h5;
-	private int s5;
-	private int v5;
-	private int h6;
-	private int s6;
-	private int v6;
-	private int h7;
-	private int s7;
-	private int v7;
-	private int h8;
-	private int s8;
-	private int v8;
-	private int h9;
-	private int s9;
-	private int v9;
-	private int h10;
-	private int s10;
-	private int v10;
-	private int h11;
-	private int s11;
-	private int v11;
-	private int h12;
-	private int s12;
-	private int v12;
-	private int h13;
-	private int s13;
-	private int v13;
-	private int h14;
-	private int s14;
-	private int v14;
-	private int h15;
-	private int s15;
-	private int v15;
-	private int h16;
-	private int s16;
-	private int v16;
+	private float h1;
+	private float s1;
+	private float v1;
+	private float h2;
+	private float s2;
+	private float v2;
+	private float h3;
+	private float s3;
+	private float v3;
+	private float h4;
+	private float s4;
+	private float v4;
+	private float h5;
+	private float s5;
+	private float v5;
+	private float h6;
+	private float s6;
+	private float v6;
+	private float h7;
+	private float s7;
+	private float v7;
+	private float h8;
+	private float s8;
+	private float v8;
+	private float h9;
+	private float s9;
+	private float v9;
+	private float h10;
+	private float s10;
+	private float v10;
+	private float h11;
+	private float s11;
+	private float v11;
+	private float h12;
+	private float s12;
+	private float v12;
+	private float h13;
+	private float s13;
+	private float v13;
+	private float h14;
+	private float s14;
+	private float v14;
+	private float h15;
+	private float s15;
+	private float v15;
+	private float h16;
+	private float s16;
+	private float v16;
 
 	public int getId() {
 		return id;
@@ -75,387 +75,387 @@ public class FingerPrint4x4 implements Serializable {
 		this.imageId = imageId;
 	}
 
-	public int getH1() {
+	public float getH1() {
 		return h1;
 	}
 
-	public void setH1(int h1) {
+	public void setH1(float h1) {
 		this.h1 = h1;
 	}
 
-	public int getS1() {
+	public float getS1() {
 		return s1;
 	}
 
-	public void setS1(int s1) {
+	public void setS1(float s1) {
 		this.s1 = s1;
 	}
 
-	public int getV1() {
+	public float getV1() {
 		return v1;
 	}
 
-	public void setV1(int v1) {
+	public void setV1(float v1) {
 		this.v1 = v1;
 	}
 
-	public int getH2() {
+	public float getH2() {
 		return h2;
 	}
 
-	public void setH2(int h2) {
+	public void setH2(float h2) {
 		this.h2 = h2;
 	}
 
-	public int getS2() {
+	public float getS2() {
 		return s2;
 	}
 
-	public void setS2(int s2) {
+	public void setS2(float s2) {
 		this.s2 = s2;
 	}
 
-	public int getV2() {
+	public float getV2() {
 		return v2;
 	}
 
-	public void setV2(int v2) {
+	public void setV2(float v2) {
 		this.v2 = v2;
 	}
 
-	public int getH3() {
+	public float getH3() {
 		return h3;
 	}
 
-	public void setH3(int h3) {
+	public void setH3(float h3) {
 		this.h3 = h3;
 	}
 
-	public int getS3() {
+	public float getS3() {
 		return s3;
 	}
 
-	public void setS3(int s3) {
+	public void setS3(float s3) {
 		this.s3 = s3;
 	}
 
-	public int getV3() {
+	public float getV3() {
 		return v3;
 	}
 
-	public void setV3(int v3) {
+	public void setV3(float v3) {
 		this.v3 = v3;
 	}
 
-	public int getH4() {
+	public float getH4() {
 		return h4;
 	}
 
-	public void setH4(int h4) {
+	public void setH4(float h4) {
 		this.h4 = h4;
 	}
 
-	public int getS4() {
+	public float getS4() {
 		return s4;
 	}
 
-	public void setS4(int s4) {
+	public void setS4(float s4) {
 		this.s4 = s4;
 	}
 
-	public int getV4() {
+	public float getV4() {
 		return v4;
 	}
 
-	public void setV4(int v4) {
+	public void setV4(float v4) {
 		this.v4 = v4;
 	}
 
-	public int getH5() {
+	public float getH5() {
 		return h5;
 	}
 
-	public void setH5(int h5) {
+	public void setH5(float h5) {
 		this.h5 = h5;
 	}
 
-	public int getS5() {
+	public float getS5() {
 		return s5;
 	}
 
-	public void setS5(int s5) {
+	public void setS5(float s5) {
 		this.s5 = s5;
 	}
 
-	public int getV5() {
+	public float getV5() {
 		return v5;
 	}
 
-	public void setV5(int v5) {
+	public void setV5(float v5) {
 		this.v5 = v5;
 	}
 
-	public int getH6() {
+	public float getH6() {
 		return h6;
 	}
 
-	public void setH6(int h6) {
+	public void setH6(float h6) {
 		this.h6 = h6;
 	}
 
-	public int getS6() {
+	public float getS6() {
 		return s6;
 	}
 
-	public void setS6(int s6) {
+	public void setS6(float s6) {
 		this.s6 = s6;
 	}
 
-	public int getV6() {
+	public float getV6() {
 		return v6;
 	}
 
-	public void setV6(int v6) {
+	public void setV6(float v6) {
 		this.v6 = v6;
 	}
 
-	public int getH7() {
+	public float getH7() {
 		return h7;
 	}
 
-	public void setH7(int h7) {
+	public void setH7(float h7) {
 		this.h7 = h7;
 	}
 
-	public int getS7() {
+	public float getS7() {
 		return s7;
 	}
 
-	public void setS7(int s7) {
+	public void setS7(float s7) {
 		this.s7 = s7;
 	}
 
-	public int getV7() {
+	public float getV7() {
 		return v7;
 	}
 
-	public void setV7(int v7) {
+	public void setV7(float v7) {
 		this.v7 = v7;
 	}
 
-	public int getH8() {
+	public float getH8() {
 		return h8;
 	}
 
-	public void setH8(int h8) {
+	public void setH8(float h8) {
 		this.h8 = h8;
 	}
 
-	public int getS8() {
+	public float getS8() {
 		return s8;
 	}
 
-	public void setS8(int s8) {
+	public void setS8(float s8) {
 		this.s8 = s8;
 	}
 
-	public int getV8() {
+	public float getV8() {
 		return v8;
 	}
 
-	public void setV8(int v8) {
+	public void setV8(float v8) {
 		this.v8 = v8;
 	}
 
-	public int getH9() {
+	public float getH9() {
 		return h9;
 	}
 
-	public void setH9(int h9) {
+	public void setH9(float h9) {
 		this.h9 = h9;
 	}
 
-	public int getS9() {
+	public float getS9() {
 		return s9;
 	}
 
-	public void setS9(int s9) {
+	public void setS9(float s9) {
 		this.s9 = s9;
 	}
 
-	public int getV9() {
+	public float getV9() {
 		return v9;
 	}
 
-	public void setV9(int v9) {
+	public void setV9(float v9) {
 		this.v9 = v9;
 	}
 
-	public int getH10() {
+	public float getH10() {
 		return h10;
 	}
 
-	public void setH10(int h10) {
+	public void setH10(float h10) {
 		this.h10 = h10;
 	}
 
-	public int getS10() {
+	public float getS10() {
 		return s10;
 	}
 
-	public void setS10(int s10) {
+	public void setS10(float s10) {
 		this.s10 = s10;
 	}
 
-	public int getV10() {
+	public float getV10() {
 		return v10;
 	}
 
-	public void setV10(int v10) {
+	public void setV10(float v10) {
 		this.v10 = v10;
 	}
 
-	public int getH11() {
+	public float getH11() {
 		return h11;
 	}
 
-	public void setH11(int h11) {
+	public void setH11(float h11) {
 		this.h11 = h11;
 	}
 
-	public int getS11() {
+	public float getS11() {
 		return s11;
 	}
 
-	public void setS11(int s11) {
+	public void setS11(float s11) {
 		this.s11 = s11;
 	}
 
-	public int getV11() {
+	public float getV11() {
 		return v11;
 	}
 
-	public void setV11(int v11) {
+	public void setV11(float v11) {
 		this.v11 = v11;
 	}
 
-	public int getH12() {
+	public float getH12() {
 		return h12;
 	}
 
-	public void setH12(int h12) {
+	public void setH12(float h12) {
 		this.h12 = h12;
 	}
 
-	public int getS12() {
+	public float getS12() {
 		return s12;
 	}
 
-	public void setS12(int s12) {
+	public void setS12(float s12) {
 		this.s12 = s12;
 	}
 
-	public int getV12() {
+	public float getV12() {
 		return v12;
 	}
 
-	public void setV12(int v12) {
+	public void setV12(float v12) {
 		this.v12 = v12;
 	}
 
-	public int getH13() {
+	public float getH13() {
 		return h13;
 	}
 
-	public void setH13(int h13) {
+	public void setH13(float h13) {
 		this.h13 = h13;
 	}
 
-	public int getS13() {
+	public float getS13() {
 		return s13;
 	}
 
-	public void setS13(int s13) {
+	public void setS13(float s13) {
 		this.s13 = s13;
 	}
 
-	public int getV13() {
+	public float getV13() {
 		return v13;
 	}
 
-	public void setV13(int v13) {
+	public void setV13(float v13) {
 		this.v13 = v13;
 	}
 
-	public int getH14() {
+	public float getH14() {
 		return h14;
 	}
 
-	public void setH14(int h14) {
+	public void setH14(float h14) {
 		this.h14 = h14;
 	}
 
-	public int getS14() {
+	public float getS14() {
 		return s14;
 	}
 
-	public void setS14(int s14) {
+	public void setS14(float s14) {
 		this.s14 = s14;
 	}
 
-	public int getV14() {
+	public float getV14() {
 		return v14;
 	}
 
-	public void setV14(int v14) {
+	public void setV14(float v14) {
 		this.v14 = v14;
 	}
 
-	public int getH15() {
+	public float getH15() {
 		return h15;
 	}
 
-	public void setH15(int h15) {
+	public void setH15(float h15) {
 		this.h15 = h15;
 	}
 
-	public int getS15() {
+	public float getS15() {
 		return s15;
 	}
 
-	public void setS15(int s15) {
+	public void setS15(float s15) {
 		this.s15 = s15;
 	}
 
-	public int getV15() {
+	public float getV15() {
 		return v15;
 	}
 
-	public void setV15(int v15) {
+	public void setV15(float v15) {
 		this.v15 = v15;
 	}
 
-	public int getH16() {
+	public float getH16() {
 		return h16;
 	}
 
-	public void setH16(int h16) {
+	public void setH16(float h16) {
 		this.h16 = h16;
 	}
 
-	public int getS16() {
+	public float getS16() {
 		return s16;
 	}
 
-	public void setS16(int s16) {
+	public void setS16(float s16) {
 		this.s16 = s16;
 	}
 
-	public int getV16() {
+	public float getV16() {
 		return v16;
 	}
 
-	public void setV16(int v16) {
+	public void setV16(float v16) {
 		this.v16 = v16;
 	}
